requirements.txt
flask
azure
jinja
requests
